make it run with python3 fix PEP issues rename Chassis
