+++ genetic-car-starters/genetic-car-starter-python/setup.cfg
@@ -1,4 +1,4 @@
-[pytest]
+[tool:pytest]
 pep8ignore =
     *.py E303 E501
     install/install_conf.py E303 E501 E225
